fix(permission_check): Fix permission check when user have multiple namespace_roles
The issue happens when user has multiple namespace role. Python uses copy
by reference. So that in the first iteration of namespace, all the
{namespace} placeholder will be replaced with the first namespace, and it
will be stored in cache. Thus, it will prevent the second namespace to
replace the {namespace} placeholder.

To fix that, we use deepcopy before replacing the placeholder, so that the
cache will not be updated in _apply_user_permission_resource_values().