fix(permission_check): Fix permission check when user have multiple namespace_roles (#28)
* fix(permission_check): Fix permission check when user have multiple namespace_roles

The issue happens when user has multiple namespace role. Python uses copy
by reference. So that in the first iteration of namespace, all the
{namespace} placeholder will be replaced with the first namespace, and it
will be stored in cache. Thus, it will prevent the second namespace to
replace the {namespace} placeholder.

To fix that, we use deepcopy before replacing the placeholder, so that the
cache will not be updated in _apply_user_permission_resource_values().

* fix(client): Disable default verify aud by pyjwt.

IAM-Python-SDK have custom audience validation, so the default validation by pyjwt should be disabled.

* chore(logging): Reformat logging.

* Bump version: 1.3.1 → 1.3.2

Co-authored-by: Nur <nur.faizin@accelbyte.net>